refactor: rename transaction send ipfs
