+++ src/domains/transaction/pages/TransactionSendIPFS/TransactionSendIPFS.stories.tsx
@@ -1,11 +1,11 @@
 import React from "react";
 
-import { TransactionSendModalIPFS } from "./TransactionSendModalIPFS";
+import { TransactionSendIPFS } from "./TransactionSendIPFS";
 
 export default { title: "Transactions / Pages / Send IPFS Modal" };
 
 export const Default = () => (
 	<div className="mt-15">
-		<TransactionSendModalIPFS />
+		<TransactionSendIPFS />
 	</div>
 );
