+++ src/domains/transaction/pages/TransactionSendIPFS/index.tsx
@@ -1,0 +1,1 @@
+export * from "/TransactionSendIPFS";
