--- src/domains/transaction/pages/TransactionSendModalIPFS/index.tsx
@@ -1,1 +1,0 @@
-export * from "/TransactionSendModalIPFS";
